+++ src/rust/target/cxxbridge/lbug/src/ffi.rs.cc
@@ -1,0 +1,1 @@
+../../../debug/build/lbug-f3c1fa6a9c6f1d97/out/cxxbridge/sources/lbug/src/ffi.rs.cc